fix: corrected logic for generating document file name
Refs: #524
