+++ src/test/java/ch/sbb/polarion/extension/pdf_exporter/service/PdfExporterPolarionServiceTest.java
@@ -3,42 +3,28 @@ package ch.sbb.polarion.extension.pdf_exporter.service;
 import ch.sbb.polarion.extension.generic.settings.NamedSettingsRegistry;
 import ch.sbb.polarion.extension.generic.settings.SettingId;
 import ch.sbb.polarion.extension.generic.settings.SettingName;
-import ch.sbb.polarion.extension.generic.test_extensions.PlatformContextMockExtension;
-import ch.sbb.polarion.extension.generic.test_extensions.TransactionalExecutorExtension;
 import ch.sbb.polarion.extension.generic.util.ScopeUtils;
 import ch.sbb.polarion.extension.pdf_exporter.rest.model.attachments.TestRunAttachment;
-import ch.sbb.polarion.extension.pdf_exporter.rest.model.collections.DocumentCollectionEntry;
 import ch.sbb.polarion.extension.pdf_exporter.rest.model.settings.stylepackage.DocIdentifier;
 import ch.sbb.polarion.extension.pdf_exporter.rest.model.settings.stylepackage.StylePackageModel;
 import ch.sbb.polarion.extension.pdf_exporter.rest.model.settings.stylepackage.StylePackageWeightInfo;
 import ch.sbb.polarion.extension.pdf_exporter.settings.StylePackageSettings;
-import ch.sbb.polarion.extension.pdf_exporter.util.DocumentFileNameHelper;
 import com.polarion.alm.projects.IProjectService;
-import com.polarion.alm.shared.api.model.baselinecollection.BaselineCollection;
-import com.polarion.alm.shared.api.model.baselinecollection.BaselineCollectionReference;
-import com.polarion.alm.shared.api.transaction.ReadOnlyTransaction;
 import com.polarion.alm.tracker.ITestManagementService;
 import com.polarion.alm.tracker.ITrackerService;
-import com.polarion.alm.tracker.model.IModule;
 import com.polarion.alm.tracker.model.ITestRecord;
 import com.polarion.alm.tracker.model.ITestRun;
 import com.polarion.alm.tracker.model.ITestRunAttachment;
 import com.polarion.alm.tracker.model.ITestStepResult;
 import com.polarion.alm.tracker.model.ITrackerProject;
 import com.polarion.alm.tracker.model.IWorkItem;
-import com.polarion.alm.tracker.model.baselinecollection.IBaselineCollection;
-import com.polarion.alm.tracker.model.baselinecollection.IBaselineCollectionElement;
 import com.polarion.platform.IPlatformService;
 import com.polarion.platform.persistence.spi.PObjectList;
 import com.polarion.platform.security.ISecurityService;
 import com.polarion.platform.service.repository.IRepositoryService;
-import com.polarion.subterra.base.location.ILocation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.MockedConstruction;
 import org.mockito.MockedStatic;
-import org.mockito.Mockito;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -47,13 +33,11 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-@ExtendWith({PlatformContextMockExtension.class, TransactionalExecutorExtension.class})
 class PdfExporterPolarionServiceTest {
 
     private final ITrackerService trackerService = mock(ITrackerService.class);
     private final ITestManagementService testManagementService = mock(ITestManagementService.class);
     private StylePackageSettings stylePackageSettings;
-    private final DocumentFileNameHelper documentFileNameHelper = mock(DocumentFileNameHelper.class);
 
     private final PdfExporterPolarionService service = new PdfExporterPolarionService(
             trackerService,
@@ -61,8 +45,7 @@ class PdfExporterPolarionServiceTest {
             mock(ISecurityService.class),
             mock(IPlatformService.class),
             mock(IRepositoryService.class),
-            testManagementService,
-            documentFileNameHelper
+            testManagementService
     );
 
     @BeforeEach
@@ -71,7 +54,6 @@ class PdfExporterPolarionServiceTest {
         when(stylePackageSettings.getFeatureName()).thenReturn("style-package");
         NamedSettingsRegistry.INSTANCE.getAll().clear();
         NamedSettingsRegistry.INSTANCE.register(List.of(stylePackageSettings));
-        when(documentFileNameHelper.getDocumentFileName(any())).thenReturn("testDocumentName.pdf");
     }
 
     @Test
@@ -355,97 +337,4 @@ class PdfExporterPolarionServiceTest {
 
         assertThrows(IllegalArgumentException.class, () -> service.getTestRunAttachment("testProjectId", "testTestRunId", "nonExistingAttachmentId", null));
     }
-
-    @Test
-    @SuppressWarnings("unused")
-    void testGetDocumentsFromCollection() {
-        String projectId = "testProjectId";
-        String collectionId = "testCollectionId";
-
-        IBaselineCollection mockCollection = mock(IBaselineCollection.class);
-        BaselineCollectionReference mockBaselineCollectionReference = mock(BaselineCollectionReference.class);
-
-        IModule mockModule1 = mock(IModule.class);
-        IModule mockModule2 = mock(IModule.class);
-        IModule mockModule3 = mock(IModule.class);
-
-        when(mockModule1.getProjectId()).thenReturn(projectId);
-        when(mockModule1.getModuleFolder()).thenReturn("space 1");
-        when(mockModule1.getModuleName()).thenReturn("test Module1");
-        ILocation mockLocation1 = mock(ILocation.class);
-        when(mockModule1.getModuleLocation()).thenReturn(mockLocation1);
-        when(mockLocation1.getLocationPath()).thenReturn("space 1");
-        when(mockModule1.getRevision()).thenReturn("1");
-
-        when(mockModule2.getProjectId()).thenReturn(projectId);
-        when(mockModule2.getModuleFolder()).thenReturn("_default");
-        when(mockModule2.getModuleName()).thenReturn("test Module2");
-        ILocation mockLocation2 = mock(ILocation.class);
-        when(mockModule2.getModuleLocation()).thenReturn(mockLocation2);
-        when(mockLocation2.getLocationPath()).thenReturn("space 2");
-        when(mockModule2.getRevision()).thenReturn("2");
-
-        when(mockModule3.getProjectId()).thenReturn(projectId);
-        when(mockModule3.getModuleFolder()).thenReturn("upstream_space");
-        when(mockModule3.getModuleName()).thenReturn("upstream Module");
-        ILocation mockLocation3 = mock(ILocation.class);
-        when(mockModule3.getModuleLocation()).thenReturn(mockLocation3);
-        when(mockLocation3.getLocationPath()).thenReturn("space 3");
-        when(mockModule3.getRevision()).thenReturn("3");
-
-        IBaselineCollectionElement mockElement1 = mock(IBaselineCollectionElement.class);
-        IBaselineCollectionElement mockElement2 = mock(IBaselineCollectionElement.class);
-        IBaselineCollectionElement mockUpstreamElement = mock(IBaselineCollectionElement.class);
-
-        when(mockElement1.getObjectWithRevision()).thenReturn(mockModule1);
-        when(mockElement2.getObjectWithRevision()).thenReturn(mockModule2);
-        when(mockUpstreamElement.getObjectWithRevision()).thenReturn(mockModule3);
-
-        IBaselineCollection mockUpstreamCollection = mock(IBaselineCollection.class);
-        when(mockUpstreamCollection.getElements()).thenReturn(List.of(mockUpstreamElement));
-        when(mockCollection.getElements()).thenReturn(List.of(mockElement1, mockElement2));
-        when(mockCollection.getUpstreamCollections()).thenReturn(List.of(mockUpstreamCollection));
-
-        BaselineCollection baselineCollection = mock(BaselineCollection.class);
-        when(mockBaselineCollectionReference.get(Mockito.any())).thenReturn(baselineCollection);
-        when(baselineCollection.getOldApi()).thenReturn(mockCollection);
-
-        try (MockedConstruction<BaselineCollectionReference> mockedStaticReference = mockConstruction(BaselineCollectionReference.class, (mock, context) -> {
-            when(mock.get(Mockito.any())).thenReturn(baselineCollection);
-            when(mock.getWithRevision(Mockito.anyString())).thenReturn(mock);
-        })) {
-            List<DocumentCollectionEntry> result = service.getDocumentsFromCollection(projectId, collectionId, null, mock(ReadOnlyTransaction.class));
-
-            assertNotNull(result);
-            assertEquals(3, result.size());
-            assertEquals("space 1", result.get(0).getSpaceId());
-            assertEquals("test Module1", result.get(0).getDocumentName());
-            assertEquals("1", result.get(0).getRevision());
-
-            assertEquals("_default", result.get(1).getSpaceId());
-            assertEquals("test Module2", result.get(1).getDocumentName());
-            assertEquals("2", result.get(1).getRevision());
-
-            assertEquals("upstream_space", result.get(2).getSpaceId());
-            assertEquals("upstream Module", result.get(2).getDocumentName());
-            assertEquals("3", result.get(2).getRevision());
-
-            List<DocumentCollectionEntry> resultWithRevision = service.getDocumentsFromCollection(projectId, collectionId, "1234", mock(ReadOnlyTransaction.class));
-
-            assertNotNull(resultWithRevision);
-            assertEquals(3, resultWithRevision.size());
-            assertEquals("space 1", resultWithRevision.get(0).getSpaceId());
-            assertEquals("test Module1", resultWithRevision.get(0).getDocumentName());
-            assertEquals("1", resultWithRevision.get(0).getRevision());
-
-            assertEquals("_default", resultWithRevision.get(1).getSpaceId());
-            assertEquals("test Module2", resultWithRevision.get(1).getDocumentName());
-            assertEquals("2", resultWithRevision.get(1).getRevision());
-
-            assertEquals("upstream_space", resultWithRevision.get(2).getSpaceId());
-            assertEquals("upstream Module", resultWithRevision.get(2).getDocumentName());
-            assertEquals("3", resultWithRevision.get(2).getRevision());
-        }
-    }
-
 }
