+++ src/test/java/ch/sbb/polarion/extension/pdf_exporter/weasyprint/PdfConverterWeasyPrintTest.java
@@ -20,7 +20,6 @@ import ch.sbb.polarion.extension.pdf_exporter.settings.CssSettings;
 import ch.sbb.polarion.extension.pdf_exporter.settings.HeaderFooterSettings;
 import ch.sbb.polarion.extension.pdf_exporter.settings.LocalizationSettings;
 import ch.sbb.polarion.extension.pdf_exporter.util.DocumentDataFactory;
-import ch.sbb.polarion.extension.pdf_exporter.util.DocumentFileNameHelper;
 import ch.sbb.polarion.extension.pdf_exporter.util.FileResourceProvider;
 import ch.sbb.polarion.extension.pdf_exporter.util.HtmlProcessor;
 import ch.sbb.polarion.extension.pdf_exporter.util.MediaUtils;
@@ -109,8 +108,7 @@ class PdfConverterWeasyPrintTest extends BaseWeasyPrintTest {
                 mock(ISecurityService.class),
                 mock(IPlatformService.class),
                 mock(IRepositoryService.class),
-                mock(ITestManagementService.class),
-                mock(DocumentFileNameHelper.class)
+                mock(ITestManagementService.class)
         );
 
         IProject project = mock(IProject.class);
